requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.57.0
python-multipart==0.0.18
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
